+++ changelog.md
@@ -1,6 +1,6 @@
 # graphql-react changelog
 
-## Next
+## 4.0.1
 
 ### Patch